+++ src/lib/data/workExperienceData.js
@@ -88,45 +88,3 @@ export const workData = [
         date: '2016 jul'
     }
 ];
-
-export const internshipData = [
-    {
-        title: 'App4mation',
-        subtitle: 'Master thesis project',
-        description: 'Story level risk assessment in agile\n' +
-            'software development using\n' +
-            'machine-learning',
-        date: '2022 May',
-        image: app4mationLogo,
-        attachments: [{
-            title: 'Master Thesis.pdf',
-            image: '/assets/file.svg',
-            url: '/assets/Thesis_Risk_assessment_in_agile_software_development.pdf'
-        }]
-    },
-    {
-        title: 'Plat4mation',
-        subtitle: 'Bachelor graduation project',
-        description: 'Build a set of tools to make local development on a ServiceNow instance easier. Researching which tools (Gulp, Grunt, etc) where the most fitting for Plat4mation\'s workflow.',
-        date: '2018 Feb',
-        image: plat4mationLogo,
-        attachments: [{
-            title: 'Bachelor Thesis.pdf',
-            image: '/assets/file.svg',
-            url: '/assets/Third_party_tool_integration_in_a_service_based_cloud_ecosystem.pdf'
-        }]
-    },
-    {
-        title: 'Microsoft',
-        subtitle: 'Eduneric gamemaker',
-        description: 'With a team of international students I worked on a C# UWP application, which generated browser based physicis games to explain physics concepts.',
-        date: '2015 Sep',
-        image: getImage('ms_logo')
-    },
-    {
-        title: 'Hostnet Bv',
-        subtitle: 'PHP Configuration validator',
-        description: 'A system which validated dependency injection configuration files, in the form of a PHP cli interface. ',
-        date: '2015 Jan'
-    },
-];
